:art: Rebase and restructure files
