+++ app/src/main/java/com/example/freightcrayt/activities/MainActivity.java
@@ -1,19 +1,23 @@
-package com.example.freightcrayt;
+package com.example.freightcrayt.activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.viewpager.widget.ViewPager;
 import androidx.viewpager2.widget.ViewPager2;
 
 import android.os.Bundle;
 
+import com.example.freightcrayt.utils.IntentHelper;
+import com.example.freightcrayt.R;
+import com.example.freightcrayt.adapters.TabAdapter;
+import com.example.freightcrayt.fragments.collaborationCategory;
+import com.example.freightcrayt.fragments.personalCategory;
 import com.google.android.material.tabs.TabLayout;
 import com.google.android.material.tabs.TabLayoutMediator;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 
-public class Login extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity {
 
+    // pagers for logic
     private TabAdapter adapter;
     private TabLayout tabLayout;
     private ViewPager2 viewPager;
@@ -21,23 +25,31 @@ public class Login extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
 
-        viewPager = (ViewPager2) findViewById(R.id.loginViewPager);
-        tabLayout = (TabLayout) findViewById(R.id.loginTabLayout);
+        // redirect to login screen if not logged in
+        FirebaseAuth mAuth = FirebaseAuth.getInstance();
+        if(mAuth.getCurrentUser().getUid().isEmpty()){
+            IntentHelper.openIntent(MainActivity.this, "Not logged in", Login.class);
+            return;
+        }
+
+        setContentView(R.layout.activity_main);
+
+        viewPager = (ViewPager2) findViewById(R.id.mainViewPager);
+        tabLayout = (TabLayout) findViewById(R.id.mainTabLayout);
 
         adapter = new TabAdapter(getSupportFragmentManager(), getLifecycle());
-        adapter.addFragment(LoginFragment.newInstance());
-        adapter.addFragment(RegisterFragment.newInstance());
+        adapter.addFragment(personalCategory.newInstance());
+        adapter.addFragment(collaborationCategory.newInstance());
 
         viewPager.setOrientation(ViewPager2.ORIENTATION_HORIZONTAL);
         viewPager.setAdapter(adapter);
         new TabLayoutMediator(tabLayout, viewPager, new TabLayoutMediator.TabConfigurationStrategy() {
             @Override public void onConfigureTab(@NonNull TabLayout.Tab tab, int position) {
                 if (position == 0) {
-                    tab.setText("Login");
+                    tab.setText("Personal");
                 } else {
-                    tab.setText("Join");
+                    tab.setText("Collaboration");
                 }
             }
         }).attach();
