+++ app/src/main/java/com/example/freightcrayt/adapters/TabAdapter.java
@@ -1,4 +1,4 @@
-package com.example.freightcrayt;
+package com.example.freightcrayt.adapters;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
