+++ app/src/main/java/com/example/freightcrayt/fragments/LoginFragment.java
@@ -1,28 +1,26 @@
-package com.example.freightcrayt;
+package com.example.freightcrayt.fragments;
 
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.Toast;
 
+import com.example.freightcrayt.utils.DataHelper;
+import com.example.freightcrayt.utils.IntentHelper;
+import com.example.freightcrayt.R;
+import com.example.freightcrayt.activities.MainActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
-import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textfield.TextInputLayout;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthException;
-import com.google.firebase.auth.FirebaseUser;
-
-import java.io.Console;
 
 public class LoginFragment extends Fragment {
