+++ app/src/main/java/com/example/freightcrayt/fragments/collaborationCategory.java
@@ -1,4 +1,4 @@
-package com.example.freightcrayt;
+package com.example.freightcrayt.fragments;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.freightcrayt.R;
 
 public class collaborationCategory extends Fragment {
