+++ app/src/main/java/com/example/freightcrayt/utils/IntentHelper.java
@@ -1,4 +1,4 @@
-package com.example.freightcrayt;
+package com.example.freightcrayt.utils;
 
 import android.content.Context;
 import android.content.Intent;
