+++ app/src/main/res/layout/activity_add_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddItem">
+    tools:context=".activities.AddItem">
 
     <LinearLayout
         android:layout_width="395dp"
