+++ app/src/main/res/layout/fragment_collaboration_category.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".collaborationCategory">
+    tools:context=".fragments.collaborationCategory">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
