+++ app/src/main/res/layout/fragment_personal_catagory.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".personalCategory">
+    tools:context=".fragments.personalCategory">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/textInputLayout"
