+++ app/src/main/res/layout/fragment_register.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisterFragment">
+    tools:context=".fragments.RegisterFragment">
 
     <TextView
         android:id="@+id/textTitleSingUp"
